UserTarget

+++ KidSteps/KidSteps/Models/Family.cs
@@ -13,7 +13,8 @@ namespace KidSteps.Models
             Members = new List<FamilyMember>();
         }
 
-        public long Id { get; set; }
+        public int Id { get; set; }
+
         [Required]
         public string Name { get; set; }
